refactor: moved some of the "game" reducers to user.reducer.js for clarity
