+++ src/redux/reducers/user.reducer.js
@@ -1,3 +1,5 @@
+import { combineReducers } from "redux";
+
 const userReducer = (state = {}, { type, payload }) => {
   return {
     'SET_USER': payload,
@@ -5,5 +7,30 @@ const userReducer = (state = {}, { type, payload }) => {
   }[type] || state
 };
 
+const userWishlist = (state = [], { type, payload }) => {
+  return {
+    'USER/SET_WISHLIST': payload,
+    'USER/CLEAR_WISHLIST': []
+  }[type] || state;
+}
 
-export default userReducer;
+const userIgnorelist = (state = [], { type, payload }) => {
+  return {
+    'USER/SET_IGNORELIST': payload,
+    'USER/CLEAR_IGNORELIST': []
+  }[type] || state;
+}
+
+const userPlayedList = (state = [], { type, payload }) => {
+  return {
+    'USER/SET_PLAYED': payload,
+    'USER/CLEAR_PLAYED': []
+  }[type] || state;
+}
+
+export default combineReducers({
+  userReducer,
+  userWishlist,
+  userIgnorelist,
+  userPlayedList,
+});
